+++ VpnHood.Tunneling/Messaging/HelloResponse.cs
@@ -6,15 +6,15 @@ using VpnHood.Common.Net;
 
 namespace VpnHood.Tunneling.Messaging;
 
-public class HelloSessionResponse : SessionResponse
+public class HelloResponse : SessionResponse
 {
     [JsonConstructor]
-    public HelloSessionResponse(SessionErrorCode errorCode)
+    public HelloResponse(SessionErrorCode errorCode)
         : base(errorCode)
     {
     }
 
-    public HelloSessionResponse(SessionResponse obj)
+    public HelloResponse(SessionResponse obj)
         : base(obj)
     {
     }
@@ -30,7 +30,7 @@ public class HelloSessionResponse : SessionResponse
 
     public string ServerVersion { get; set; } = default!;
     public int ServerProtocolVersion { get; set; }
-    
+    public string? ServerTokenUrl { get; set; }
     public byte[] ServerSecret { get; set; } = default!;
     public int MaxDatagramChannelCount { get; set; } 
     public bool IsIpV6Supported { get; set; }
@@ -39,5 +39,4 @@ public class HelloSessionResponse : SessionResponse
     public string? GaMeasurementId { get; init;}
     public TimeSpan RequestTimeout { get; init; } = TimeSpan.FromSeconds(60);
     public TimeSpan TcpReuseTimeout { get; init; } = TimeSpan.FromSeconds(60);
-    public string? ServerTokenUrl { get; set; }
 }